alex : Lot of work on 3rd enemy - fx - Arena Entrance (big commit)
Former-commit-id: 73901bef4979db49b6f878f6874d3ed4a86201a2
Former-commit-id: c9cedc11c0018bd0a6653696b02d677859bdaae6

+++ Assets/Scripts/Battle Scripts/FireExplosionDelay.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class FireExplosionDelay : MonoBehaviour {
 
     public float _Speed;
-    public int _Radius = 110;
+    public float _Radius = 110;
     public float _Delay;
 
     SphereCollider _Collider;
@@ -17,7 +17,6 @@ public class FireExplosionDelay : MonoBehaviour {
         _Collider = GetComponent<SphereCollider>();
 	}
 	
-	// Update is called once per frame
 	void Update () {
         if (timer >= _Delay)
         {
